Generate (semantic) colours in Swift as class members
Move core colours into their own class but leaving them as static.
