+++ assets/ios/swift/CompoundCoreColorTokens.swift
@@ -1,0 +1,320 @@
+
+//
+// CompoundCoreColorTokens.swift
+//
+
+import SwiftUI
+
+public class CompoundCoreColorTokens {
+    public static let alphaPink1400 = Color("alphaPink1400", bundle: Bundle.module) 
+    public static let alphaPink1300 = Color("alphaPink1300", bundle: Bundle.module) 
+    public static let alphaPink1200 = Color("alphaPink1200", bundle: Bundle.module) 
+    public static let alphaPink1100 = Color("alphaPink1100", bundle: Bundle.module) 
+    public static let alphaPink1000 = Color("alphaPink1000", bundle: Bundle.module) 
+    public static let alphaPink900 = Color("alphaPink900", bundle: Bundle.module) 
+    public static let alphaPink800 = Color("alphaPink800", bundle: Bundle.module) 
+    public static let alphaPink700 = Color("alphaPink700", bundle: Bundle.module) 
+    public static let alphaPink600 = Color("alphaPink600", bundle: Bundle.module) 
+    public static let alphaPink500 = Color("alphaPink500", bundle: Bundle.module) 
+    public static let alphaPink400 = Color("alphaPink400", bundle: Bundle.module) 
+    public static let alphaPink300 = Color("alphaPink300", bundle: Bundle.module) 
+    public static let alphaPink200 = Color("alphaPink200", bundle: Bundle.module) 
+    public static let alphaPink100 = Color("alphaPink100", bundle: Bundle.module) 
+    public static let alphaFuchsia1400 = Color("alphaFuchsia1400", bundle: Bundle.module) 
+    public static let alphaFuchsia1300 = Color("alphaFuchsia1300", bundle: Bundle.module) 
+    public static let alphaFuchsia1200 = Color("alphaFuchsia1200", bundle: Bundle.module) 
+    public static let alphaFuchsia1100 = Color("alphaFuchsia1100", bundle: Bundle.module) 
+    public static let alphaFuchsia1000 = Color("alphaFuchsia1000", bundle: Bundle.module) 
+    public static let alphaFuchsia900 = Color("alphaFuchsia900", bundle: Bundle.module) 
+    public static let alphaFuchsia800 = Color("alphaFuchsia800", bundle: Bundle.module) 
+    public static let alphaFuchsia700 = Color("alphaFuchsia700", bundle: Bundle.module) 
+    public static let alphaFuchsia600 = Color("alphaFuchsia600", bundle: Bundle.module) 
+    public static let alphaFuchsia500 = Color("alphaFuchsia500", bundle: Bundle.module) 
+    public static let alphaFuchsia400 = Color("alphaFuchsia400", bundle: Bundle.module) 
+    public static let alphaFuchsia300 = Color("alphaFuchsia300", bundle: Bundle.module) 
+    public static let alphaFuchsia200 = Color("alphaFuchsia200", bundle: Bundle.module) 
+    public static let alphaFuchsia100 = Color("alphaFuchsia100", bundle: Bundle.module) 
+    public static let alphaPurple1400 = Color("alphaPurple1400", bundle: Bundle.module) 
+    public static let alphaPurple1300 = Color("alphaPurple1300", bundle: Bundle.module) 
+    public static let alphaPurple1200 = Color("alphaPurple1200", bundle: Bundle.module) 
+    public static let alphaPurple1100 = Color("alphaPurple1100", bundle: Bundle.module) 
+    public static let alphaPurple1000 = Color("alphaPurple1000", bundle: Bundle.module) 
+    public static let alphaPurple900 = Color("alphaPurple900", bundle: Bundle.module) 
+    public static let alphaPurple800 = Color("alphaPurple800", bundle: Bundle.module) 
+    public static let alphaPurple700 = Color("alphaPurple700", bundle: Bundle.module) 
+    public static let alphaPurple600 = Color("alphaPurple600", bundle: Bundle.module) 
+    public static let alphaPurple500 = Color("alphaPurple500", bundle: Bundle.module) 
+    public static let alphaPurple400 = Color("alphaPurple400", bundle: Bundle.module) 
+    public static let alphaPurple300 = Color("alphaPurple300", bundle: Bundle.module) 
+    public static let alphaPurple200 = Color("alphaPurple200", bundle: Bundle.module) 
+    public static let alphaPurple100 = Color("alphaPurple100", bundle: Bundle.module) 
+    public static let alphaBlue1400 = Color("alphaBlue1400", bundle: Bundle.module) 
+    public static let alphaBlue1300 = Color("alphaBlue1300", bundle: Bundle.module) 
+    public static let alphaBlue1200 = Color("alphaBlue1200", bundle: Bundle.module) 
+    public static let alphaBlue1100 = Color("alphaBlue1100", bundle: Bundle.module) 
+    public static let alphaBlue1000 = Color("alphaBlue1000", bundle: Bundle.module) 
+    public static let alphaBlue900 = Color("alphaBlue900", bundle: Bundle.module) 
+    public static let alphaBlue800 = Color("alphaBlue800", bundle: Bundle.module) 
+    public static let alphaBlue700 = Color("alphaBlue700", bundle: Bundle.module) 
+    public static let alphaBlue600 = Color("alphaBlue600", bundle: Bundle.module) 
+    public static let alphaBlue500 = Color("alphaBlue500", bundle: Bundle.module) 
+    public static let alphaBlue400 = Color("alphaBlue400", bundle: Bundle.module) 
+    public static let alphaBlue300 = Color("alphaBlue300", bundle: Bundle.module) 
+    public static let alphaBlue200 = Color("alphaBlue200", bundle: Bundle.module) 
+    public static let alphaBlue100 = Color("alphaBlue100", bundle: Bundle.module) 
+    public static let alphaCyan1400 = Color("alphaCyan1400", bundle: Bundle.module) 
+    public static let alphaCyan1300 = Color("alphaCyan1300", bundle: Bundle.module) 
+    public static let alphaCyan1200 = Color("alphaCyan1200", bundle: Bundle.module) 
+    public static let alphaCyan1100 = Color("alphaCyan1100", bundle: Bundle.module) 
+    public static let alphaCyan1000 = Color("alphaCyan1000", bundle: Bundle.module) 
+    public static let alphaCyan900 = Color("alphaCyan900", bundle: Bundle.module) 
+    public static let alphaCyan800 = Color("alphaCyan800", bundle: Bundle.module) 
+    public static let alphaCyan700 = Color("alphaCyan700", bundle: Bundle.module) 
+    public static let alphaCyan600 = Color("alphaCyan600", bundle: Bundle.module) 
+    public static let alphaCyan500 = Color("alphaCyan500", bundle: Bundle.module) 
+    public static let alphaCyan400 = Color("alphaCyan400", bundle: Bundle.module) 
+    public static let alphaCyan300 = Color("alphaCyan300", bundle: Bundle.module) 
+    public static let alphaCyan200 = Color("alphaCyan200", bundle: Bundle.module) 
+    public static let alphaCyan100 = Color("alphaCyan100", bundle: Bundle.module) 
+    public static let alphaGreen1400 = Color("alphaGreen1400", bundle: Bundle.module) 
+    public static let alphaGreen1300 = Color("alphaGreen1300", bundle: Bundle.module) 
+    public static let alphaGreen1200 = Color("alphaGreen1200", bundle: Bundle.module) 
+    public static let alphaGreen1100 = Color("alphaGreen1100", bundle: Bundle.module) 
+    public static let alphaGreen1000 = Color("alphaGreen1000", bundle: Bundle.module) 
+    public static let alphaGreen900 = Color("alphaGreen900", bundle: Bundle.module) 
+    public static let alphaGreen800 = Color("alphaGreen800", bundle: Bundle.module) 
+    public static let alphaGreen700 = Color("alphaGreen700", bundle: Bundle.module) 
+    public static let alphaGreen600 = Color("alphaGreen600", bundle: Bundle.module) 
+    public static let alphaGreen500 = Color("alphaGreen500", bundle: Bundle.module) 
+    public static let alphaGreen400 = Color("alphaGreen400", bundle: Bundle.module) 
+    public static let alphaGreen300 = Color("alphaGreen300", bundle: Bundle.module) 
+    public static let alphaGreen200 = Color("alphaGreen200", bundle: Bundle.module) 
+    public static let alphaGreen100 = Color("alphaGreen100", bundle: Bundle.module) 
+    public static let alphaLime1400 = Color("alphaLime1400", bundle: Bundle.module) 
+    public static let alphaLime1300 = Color("alphaLime1300", bundle: Bundle.module) 
+    public static let alphaLime1200 = Color("alphaLime1200", bundle: Bundle.module) 
+    public static let alphaLime1100 = Color("alphaLime1100", bundle: Bundle.module) 
+    public static let alphaLime1000 = Color("alphaLime1000", bundle: Bundle.module) 
+    public static let alphaLime900 = Color("alphaLime900", bundle: Bundle.module) 
+    public static let alphaLime800 = Color("alphaLime800", bundle: Bundle.module) 
+    public static let alphaLime700 = Color("alphaLime700", bundle: Bundle.module) 
+    public static let alphaLime600 = Color("alphaLime600", bundle: Bundle.module) 
+    public static let alphaLime500 = Color("alphaLime500", bundle: Bundle.module) 
+    public static let alphaLime400 = Color("alphaLime400", bundle: Bundle.module) 
+    public static let alphaLime300 = Color("alphaLime300", bundle: Bundle.module) 
+    public static let alphaLime200 = Color("alphaLime200", bundle: Bundle.module) 
+    public static let alphaLime100 = Color("alphaLime100", bundle: Bundle.module) 
+    public static let alphaYellow1400 = Color("alphaYellow1400", bundle: Bundle.module) 
+    public static let alphaYellow1300 = Color("alphaYellow1300", bundle: Bundle.module) 
+    public static let alphaYellow1200 = Color("alphaYellow1200", bundle: Bundle.module) 
+    public static let alphaYellow1100 = Color("alphaYellow1100", bundle: Bundle.module) 
+    public static let alphaYellow1000 = Color("alphaYellow1000", bundle: Bundle.module) 
+    public static let alphaYellow900 = Color("alphaYellow900", bundle: Bundle.module) 
+    public static let alphaYellow800 = Color("alphaYellow800", bundle: Bundle.module) 
+    public static let alphaYellow700 = Color("alphaYellow700", bundle: Bundle.module) 
+    public static let alphaYellow600 = Color("alphaYellow600", bundle: Bundle.module) 
+    public static let alphaYellow500 = Color("alphaYellow500", bundle: Bundle.module) 
+    public static let alphaYellow400 = Color("alphaYellow400", bundle: Bundle.module) 
+    public static let alphaYellow300 = Color("alphaYellow300", bundle: Bundle.module) 
+    public static let alphaYellow200 = Color("alphaYellow200", bundle: Bundle.module) 
+    public static let alphaYellow100 = Color("alphaYellow100", bundle: Bundle.module) 
+    public static let alphaOrange1400 = Color("alphaOrange1400", bundle: Bundle.module) 
+    public static let alphaOrange1300 = Color("alphaOrange1300", bundle: Bundle.module) 
+    public static let alphaOrange1200 = Color("alphaOrange1200", bundle: Bundle.module) 
+    public static let alphaOrange1100 = Color("alphaOrange1100", bundle: Bundle.module) 
+    public static let alphaOrange1000 = Color("alphaOrange1000", bundle: Bundle.module) 
+    public static let alphaOrange900 = Color("alphaOrange900", bundle: Bundle.module) 
+    public static let alphaOrange800 = Color("alphaOrange800", bundle: Bundle.module) 
+    public static let alphaOrange700 = Color("alphaOrange700", bundle: Bundle.module) 
+    public static let alphaOrange600 = Color("alphaOrange600", bundle: Bundle.module) 
+    public static let alphaOrange500 = Color("alphaOrange500", bundle: Bundle.module) 
+    public static let alphaOrange400 = Color("alphaOrange400", bundle: Bundle.module) 
+    public static let alphaOrange300 = Color("alphaOrange300", bundle: Bundle.module) 
+    public static let alphaOrange200 = Color("alphaOrange200", bundle: Bundle.module) 
+    public static let alphaOrange100 = Color("alphaOrange100", bundle: Bundle.module) 
+    public static let alphaRed1400 = Color("alphaRed1400", bundle: Bundle.module) 
+    public static let alphaRed1300 = Color("alphaRed1300", bundle: Bundle.module) 
+    public static let alphaRed1200 = Color("alphaRed1200", bundle: Bundle.module) 
+    public static let alphaRed1100 = Color("alphaRed1100", bundle: Bundle.module) 
+    public static let alphaRed1000 = Color("alphaRed1000", bundle: Bundle.module) 
+    public static let alphaRed900 = Color("alphaRed900", bundle: Bundle.module) 
+    public static let alphaRed800 = Color("alphaRed800", bundle: Bundle.module) 
+    public static let alphaRed700 = Color("alphaRed700", bundle: Bundle.module) 
+    public static let alphaRed600 = Color("alphaRed600", bundle: Bundle.module) 
+    public static let alphaRed500 = Color("alphaRed500", bundle: Bundle.module) 
+    public static let alphaRed400 = Color("alphaRed400", bundle: Bundle.module) 
+    public static let alphaRed300 = Color("alphaRed300", bundle: Bundle.module) 
+    public static let alphaRed200 = Color("alphaRed200", bundle: Bundle.module) 
+    public static let alphaRed100 = Color("alphaRed100", bundle: Bundle.module) 
+    public static let alphaGray1400 = Color("alphaGray1400", bundle: Bundle.module) 
+    public static let alphaGray1300 = Color("alphaGray1300", bundle: Bundle.module) 
+    public static let alphaGray1200 = Color("alphaGray1200", bundle: Bundle.module) 
+    public static let alphaGray1100 = Color("alphaGray1100", bundle: Bundle.module) 
+    public static let alphaGray1000 = Color("alphaGray1000", bundle: Bundle.module) 
+    public static let alphaGray900 = Color("alphaGray900", bundle: Bundle.module) 
+    public static let alphaGray800 = Color("alphaGray800", bundle: Bundle.module) 
+    public static let alphaGray700 = Color("alphaGray700", bundle: Bundle.module) 
+    public static let alphaGray600 = Color("alphaGray600", bundle: Bundle.module) 
+    public static let alphaGray500 = Color("alphaGray500", bundle: Bundle.module) 
+    public static let alphaGray400 = Color("alphaGray400", bundle: Bundle.module) 
+    public static let alphaGray300 = Color("alphaGray300", bundle: Bundle.module) 
+    public static let alphaGray200 = Color("alphaGray200", bundle: Bundle.module) 
+    public static let alphaGray100 = Color("alphaGray100", bundle: Bundle.module) 
+    public static let pink1400 = Color("pink1400", bundle: Bundle.module) 
+    public static let pink1300 = Color("pink1300", bundle: Bundle.module) 
+    public static let pink1200 = Color("pink1200", bundle: Bundle.module) 
+    public static let pink1100 = Color("pink1100", bundle: Bundle.module) 
+    public static let pink1000 = Color("pink1000", bundle: Bundle.module) 
+    public static let pink900 = Color("pink900", bundle: Bundle.module) 
+    public static let pink800 = Color("pink800", bundle: Bundle.module) 
+    public static let pink700 = Color("pink700", bundle: Bundle.module) 
+    public static let pink600 = Color("pink600", bundle: Bundle.module) 
+    public static let pink500 = Color("pink500", bundle: Bundle.module) 
+    public static let pink400 = Color("pink400", bundle: Bundle.module) 
+    public static let pink300 = Color("pink300", bundle: Bundle.module) 
+    public static let pink200 = Color("pink200", bundle: Bundle.module) 
+    public static let pink100 = Color("pink100", bundle: Bundle.module) 
+    public static let fuchsia1400 = Color("fuchsia1400", bundle: Bundle.module) 
+    public static let fuchsia1300 = Color("fuchsia1300", bundle: Bundle.module) 
+    public static let fuchsia1200 = Color("fuchsia1200", bundle: Bundle.module) 
+    public static let fuchsia1100 = Color("fuchsia1100", bundle: Bundle.module) 
+    public static let fuchsia1000 = Color("fuchsia1000", bundle: Bundle.module) 
+    public static let fuchsia900 = Color("fuchsia900", bundle: Bundle.module) 
+    public static let fuchsia800 = Color("fuchsia800", bundle: Bundle.module) 
+    public static let fuchsia700 = Color("fuchsia700", bundle: Bundle.module) 
+    public static let fuchsia600 = Color("fuchsia600", bundle: Bundle.module) 
+    public static let fuchsia500 = Color("fuchsia500", bundle: Bundle.module) 
+    public static let fuchsia400 = Color("fuchsia400", bundle: Bundle.module) 
+    public static let fuchsia300 = Color("fuchsia300", bundle: Bundle.module) 
+    public static let fuchsia200 = Color("fuchsia200", bundle: Bundle.module) 
+    public static let fuchsia100 = Color("fuchsia100", bundle: Bundle.module) 
+    public static let purple1400 = Color("purple1400", bundle: Bundle.module) 
+    public static let purple1300 = Color("purple1300", bundle: Bundle.module) 
+    public static let purple1200 = Color("purple1200", bundle: Bundle.module) 
+    public static let purple1100 = Color("purple1100", bundle: Bundle.module) 
+    public static let purple1000 = Color("purple1000", bundle: Bundle.module) 
+    public static let purple900 = Color("purple900", bundle: Bundle.module) 
+    public static let purple800 = Color("purple800", bundle: Bundle.module) 
+    public static let purple700 = Color("purple700", bundle: Bundle.module) 
+    public static let purple600 = Color("purple600", bundle: Bundle.module) 
+    public static let purple500 = Color("purple500", bundle: Bundle.module) 
+    public static let purple400 = Color("purple400", bundle: Bundle.module) 
+    public static let purple300 = Color("purple300", bundle: Bundle.module) 
+    public static let purple200 = Color("purple200", bundle: Bundle.module) 
+    public static let purple100 = Color("purple100", bundle: Bundle.module) 
+    public static let blue1400 = Color("blue1400", bundle: Bundle.module) 
+    public static let blue1300 = Color("blue1300", bundle: Bundle.module) 
+    public static let blue1200 = Color("blue1200", bundle: Bundle.module) 
+    public static let blue1100 = Color("blue1100", bundle: Bundle.module) 
+    public static let blue1000 = Color("blue1000", bundle: Bundle.module) 
+    public static let blue900 = Color("blue900", bundle: Bundle.module) 
+    public static let blue800 = Color("blue800", bundle: Bundle.module) 
+    public static let blue700 = Color("blue700", bundle: Bundle.module) 
+    public static let blue600 = Color("blue600", bundle: Bundle.module) 
+    public static let blue500 = Color("blue500", bundle: Bundle.module) 
+    public static let blue400 = Color("blue400", bundle: Bundle.module) 
+    public static let blue300 = Color("blue300", bundle: Bundle.module) 
+    public static let blue200 = Color("blue200", bundle: Bundle.module) 
+    public static let blue100 = Color("blue100", bundle: Bundle.module) 
+    public static let cyan1400 = Color("cyan1400", bundle: Bundle.module) 
+    public static let cyan1300 = Color("cyan1300", bundle: Bundle.module) 
+    public static let cyan1200 = Color("cyan1200", bundle: Bundle.module) 
+    public static let cyan1100 = Color("cyan1100", bundle: Bundle.module) 
+    public static let cyan1000 = Color("cyan1000", bundle: Bundle.module) 
+    public static let cyan900 = Color("cyan900", bundle: Bundle.module) 
+    public static let cyan800 = Color("cyan800", bundle: Bundle.module) 
+    public static let cyan700 = Color("cyan700", bundle: Bundle.module) 
+    public static let cyan600 = Color("cyan600", bundle: Bundle.module) 
+    public static let cyan500 = Color("cyan500", bundle: Bundle.module) 
+    public static let cyan400 = Color("cyan400", bundle: Bundle.module) 
+    public static let cyan300 = Color("cyan300", bundle: Bundle.module) 
+    public static let cyan200 = Color("cyan200", bundle: Bundle.module) 
+    public static let cyan100 = Color("cyan100", bundle: Bundle.module) 
+    public static let green1400 = Color("green1400", bundle: Bundle.module) 
+    public static let green1300 = Color("green1300", bundle: Bundle.module) 
+    public static let green1200 = Color("green1200", bundle: Bundle.module) 
+    public static let green1100 = Color("green1100", bundle: Bundle.module) 
+    public static let green1000 = Color("green1000", bundle: Bundle.module) 
+    public static let green900 = Color("green900", bundle: Bundle.module) 
+    public static let green800 = Color("green800", bundle: Bundle.module) 
+    public static let green700 = Color("green700", bundle: Bundle.module) 
+    public static let green600 = Color("green600", bundle: Bundle.module) 
+    public static let green500 = Color("green500", bundle: Bundle.module) 
+    public static let green400 = Color("green400", bundle: Bundle.module) 
+    public static let green300 = Color("green300", bundle: Bundle.module) 
+    public static let green200 = Color("green200", bundle: Bundle.module) 
+    public static let green100 = Color("green100", bundle: Bundle.module) 
+    public static let lime1400 = Color("lime1400", bundle: Bundle.module) 
+    public static let lime1300 = Color("lime1300", bundle: Bundle.module) 
+    public static let lime1200 = Color("lime1200", bundle: Bundle.module) 
+    public static let lime1100 = Color("lime1100", bundle: Bundle.module) 
+    public static let lime1000 = Color("lime1000", bundle: Bundle.module) 
+    public static let lime900 = Color("lime900", bundle: Bundle.module) 
+    public static let lime800 = Color("lime800", bundle: Bundle.module) 
+    public static let lime700 = Color("lime700", bundle: Bundle.module) 
+    public static let lime600 = Color("lime600", bundle: Bundle.module) 
+    public static let lime500 = Color("lime500", bundle: Bundle.module) 
+    public static let lime400 = Color("lime400", bundle: Bundle.module) 
+    public static let lime300 = Color("lime300", bundle: Bundle.module) 
+    public static let lime200 = Color("lime200", bundle: Bundle.module) 
+    public static let lime100 = Color("lime100", bundle: Bundle.module) 
+    public static let yellow1400 = Color("yellow1400", bundle: Bundle.module) 
+    public static let yellow1300 = Color("yellow1300", bundle: Bundle.module) 
+    public static let yellow1200 = Color("yellow1200", bundle: Bundle.module) 
+    public static let yellow1100 = Color("yellow1100", bundle: Bundle.module) 
+    public static let yellow1000 = Color("yellow1000", bundle: Bundle.module) 
+    public static let yellow900 = Color("yellow900", bundle: Bundle.module) 
+    public static let yellow800 = Color("yellow800", bundle: Bundle.module) 
+    public static let yellow700 = Color("yellow700", bundle: Bundle.module) 
+    public static let yellow600 = Color("yellow600", bundle: Bundle.module) 
+    public static let yellow500 = Color("yellow500", bundle: Bundle.module) 
+    public static let yellow400 = Color("yellow400", bundle: Bundle.module) 
+    public static let yellow300 = Color("yellow300", bundle: Bundle.module) 
+    public static let yellow200 = Color("yellow200", bundle: Bundle.module) 
+    public static let yellow100 = Color("yellow100", bundle: Bundle.module) 
+    public static let orange1400 = Color("orange1400", bundle: Bundle.module) 
+    public static let orange1300 = Color("orange1300", bundle: Bundle.module) 
+    public static let orange1200 = Color("orange1200", bundle: Bundle.module) 
+    public static let orange1100 = Color("orange1100", bundle: Bundle.module) 
+    public static let orange1000 = Color("orange1000", bundle: Bundle.module) 
+    public static let orange900 = Color("orange900", bundle: Bundle.module) 
+    public static let orange800 = Color("orange800", bundle: Bundle.module) 
+    public static let orange700 = Color("orange700", bundle: Bundle.module) 
+    public static let orange600 = Color("orange600", bundle: Bundle.module) 
+    public static let orange500 = Color("orange500", bundle: Bundle.module) 
+    public static let orange400 = Color("orange400", bundle: Bundle.module) 
+    public static let orange300 = Color("orange300", bundle: Bundle.module) 
+    public static let orange200 = Color("orange200", bundle: Bundle.module) 
+    public static let orange100 = Color("orange100", bundle: Bundle.module) 
+    public static let red1400 = Color("red1400", bundle: Bundle.module) 
+    public static let red1300 = Color("red1300", bundle: Bundle.module) 
+    public static let red1200 = Color("red1200", bundle: Bundle.module) 
+    public static let red1100 = Color("red1100", bundle: Bundle.module) 
+    public static let red1000 = Color("red1000", bundle: Bundle.module) 
+    public static let red900 = Color("red900", bundle: Bundle.module) 
+    public static let red800 = Color("red800", bundle: Bundle.module) 
+    public static let red700 = Color("red700", bundle: Bundle.module) 
+    public static let red600 = Color("red600", bundle: Bundle.module) 
+    public static let red500 = Color("red500", bundle: Bundle.module) 
+    public static let red400 = Color("red400", bundle: Bundle.module) 
+    public static let red300 = Color("red300", bundle: Bundle.module) 
+    public static let red200 = Color("red200", bundle: Bundle.module) 
+    public static let red100 = Color("red100", bundle: Bundle.module) 
+    public static let gray1400 = Color("gray1400", bundle: Bundle.module) 
+    public static let gray1300 = Color("gray1300", bundle: Bundle.module) 
+    public static let gray1200 = Color("gray1200", bundle: Bundle.module) 
+    public static let gray1100 = Color("gray1100", bundle: Bundle.module) 
+    public static let gray1000 = Color("gray1000", bundle: Bundle.module) 
+    public static let gray900 = Color("gray900", bundle: Bundle.module) 
+    public static let gray800 = Color("gray800", bundle: Bundle.module) 
+    public static let gray700 = Color("gray700", bundle: Bundle.module) 
+    public static let gray600 = Color("gray600", bundle: Bundle.module) 
+    public static let gray500 = Color("gray500", bundle: Bundle.module) 
+    public static let gray400 = Color("gray400", bundle: Bundle.module) 
+    public static let gray300 = Color("gray300", bundle: Bundle.module) 
+    public static let gray200 = Color("gray200", bundle: Bundle.module) 
+    public static let gray100 = Color("gray100", bundle: Bundle.module) 
+    public static let themeBg = Color("themeBg", bundle: Bundle.module) 
+    public static let bgSubtleSecondaryLevel0 = themeBg
+    public static let bgCanvasDefaultLevel1 = gray300
+}
