+++ assets/ios/swift/CompoundUIColorTokens.swift
@@ -6,383 +6,73 @@
 import UIKit
 
 public class CompoundUIColorTokens {
-    public static let alphaPink1400 = UIColor(named: "alphaPink1400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPink1300 = UIColor(named: "alphaPink1300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPink1200 = UIColor(named: "alphaPink1200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPink1100 = UIColor(named: "alphaPink1100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPink1000 = UIColor(named: "alphaPink1000", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPink900 = UIColor(named: "alphaPink900", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPink800 = UIColor(named: "alphaPink800", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPink700 = UIColor(named: "alphaPink700", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPink600 = UIColor(named: "alphaPink600", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPink500 = UIColor(named: "alphaPink500", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPink400 = UIColor(named: "alphaPink400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPink300 = UIColor(named: "alphaPink300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPink200 = UIColor(named: "alphaPink200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPink100 = UIColor(named: "alphaPink100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaFuchsia1400 = UIColor(named: "alphaFuchsia1400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaFuchsia1300 = UIColor(named: "alphaFuchsia1300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaFuchsia1200 = UIColor(named: "alphaFuchsia1200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaFuchsia1100 = UIColor(named: "alphaFuchsia1100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaFuchsia1000 = UIColor(named: "alphaFuchsia1000", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaFuchsia900 = UIColor(named: "alphaFuchsia900", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaFuchsia800 = UIColor(named: "alphaFuchsia800", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaFuchsia700 = UIColor(named: "alphaFuchsia700", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaFuchsia600 = UIColor(named: "alphaFuchsia600", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaFuchsia500 = UIColor(named: "alphaFuchsia500", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaFuchsia400 = UIColor(named: "alphaFuchsia400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaFuchsia300 = UIColor(named: "alphaFuchsia300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaFuchsia200 = UIColor(named: "alphaFuchsia200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaFuchsia100 = UIColor(named: "alphaFuchsia100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPurple1400 = UIColor(named: "alphaPurple1400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPurple1300 = UIColor(named: "alphaPurple1300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPurple1200 = UIColor(named: "alphaPurple1200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPurple1100 = UIColor(named: "alphaPurple1100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPurple1000 = UIColor(named: "alphaPurple1000", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPurple900 = UIColor(named: "alphaPurple900", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPurple800 = UIColor(named: "alphaPurple800", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPurple700 = UIColor(named: "alphaPurple700", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPurple600 = UIColor(named: "alphaPurple600", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPurple500 = UIColor(named: "alphaPurple500", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPurple400 = UIColor(named: "alphaPurple400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPurple300 = UIColor(named: "alphaPurple300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPurple200 = UIColor(named: "alphaPurple200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaPurple100 = UIColor(named: "alphaPurple100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaBlue1400 = UIColor(named: "alphaBlue1400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaBlue1300 = UIColor(named: "alphaBlue1300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaBlue1200 = UIColor(named: "alphaBlue1200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaBlue1100 = UIColor(named: "alphaBlue1100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaBlue1000 = UIColor(named: "alphaBlue1000", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaBlue900 = UIColor(named: "alphaBlue900", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaBlue800 = UIColor(named: "alphaBlue800", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaBlue700 = UIColor(named: "alphaBlue700", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaBlue600 = UIColor(named: "alphaBlue600", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaBlue500 = UIColor(named: "alphaBlue500", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaBlue400 = UIColor(named: "alphaBlue400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaBlue300 = UIColor(named: "alphaBlue300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaBlue200 = UIColor(named: "alphaBlue200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaBlue100 = UIColor(named: "alphaBlue100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaCyan1400 = UIColor(named: "alphaCyan1400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaCyan1300 = UIColor(named: "alphaCyan1300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaCyan1200 = UIColor(named: "alphaCyan1200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaCyan1100 = UIColor(named: "alphaCyan1100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaCyan1000 = UIColor(named: "alphaCyan1000", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaCyan900 = UIColor(named: "alphaCyan900", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaCyan800 = UIColor(named: "alphaCyan800", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaCyan700 = UIColor(named: "alphaCyan700", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaCyan600 = UIColor(named: "alphaCyan600", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaCyan500 = UIColor(named: "alphaCyan500", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaCyan400 = UIColor(named: "alphaCyan400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaCyan300 = UIColor(named: "alphaCyan300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaCyan200 = UIColor(named: "alphaCyan200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaCyan100 = UIColor(named: "alphaCyan100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGreen1400 = UIColor(named: "alphaGreen1400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGreen1300 = UIColor(named: "alphaGreen1300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGreen1200 = UIColor(named: "alphaGreen1200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGreen1100 = UIColor(named: "alphaGreen1100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGreen1000 = UIColor(named: "alphaGreen1000", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGreen900 = UIColor(named: "alphaGreen900", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGreen800 = UIColor(named: "alphaGreen800", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGreen700 = UIColor(named: "alphaGreen700", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGreen600 = UIColor(named: "alphaGreen600", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGreen500 = UIColor(named: "alphaGreen500", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGreen400 = UIColor(named: "alphaGreen400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGreen300 = UIColor(named: "alphaGreen300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGreen200 = UIColor(named: "alphaGreen200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGreen100 = UIColor(named: "alphaGreen100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaLime1400 = UIColor(named: "alphaLime1400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaLime1300 = UIColor(named: "alphaLime1300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaLime1200 = UIColor(named: "alphaLime1200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaLime1100 = UIColor(named: "alphaLime1100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaLime1000 = UIColor(named: "alphaLime1000", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaLime900 = UIColor(named: "alphaLime900", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaLime800 = UIColor(named: "alphaLime800", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaLime700 = UIColor(named: "alphaLime700", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaLime600 = UIColor(named: "alphaLime600", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaLime500 = UIColor(named: "alphaLime500", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaLime400 = UIColor(named: "alphaLime400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaLime300 = UIColor(named: "alphaLime300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaLime200 = UIColor(named: "alphaLime200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaLime100 = UIColor(named: "alphaLime100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaYellow1400 = UIColor(named: "alphaYellow1400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaYellow1300 = UIColor(named: "alphaYellow1300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaYellow1200 = UIColor(named: "alphaYellow1200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaYellow1100 = UIColor(named: "alphaYellow1100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaYellow1000 = UIColor(named: "alphaYellow1000", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaYellow900 = UIColor(named: "alphaYellow900", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaYellow800 = UIColor(named: "alphaYellow800", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaYellow700 = UIColor(named: "alphaYellow700", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaYellow600 = UIColor(named: "alphaYellow600", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaYellow500 = UIColor(named: "alphaYellow500", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaYellow400 = UIColor(named: "alphaYellow400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaYellow300 = UIColor(named: "alphaYellow300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaYellow200 = UIColor(named: "alphaYellow200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaYellow100 = UIColor(named: "alphaYellow100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaOrange1400 = UIColor(named: "alphaOrange1400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaOrange1300 = UIColor(named: "alphaOrange1300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaOrange1200 = UIColor(named: "alphaOrange1200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaOrange1100 = UIColor(named: "alphaOrange1100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaOrange1000 = UIColor(named: "alphaOrange1000", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaOrange900 = UIColor(named: "alphaOrange900", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaOrange800 = UIColor(named: "alphaOrange800", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaOrange700 = UIColor(named: "alphaOrange700", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaOrange600 = UIColor(named: "alphaOrange600", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaOrange500 = UIColor(named: "alphaOrange500", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaOrange400 = UIColor(named: "alphaOrange400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaOrange300 = UIColor(named: "alphaOrange300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaOrange200 = UIColor(named: "alphaOrange200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaOrange100 = UIColor(named: "alphaOrange100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaRed1400 = UIColor(named: "alphaRed1400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaRed1300 = UIColor(named: "alphaRed1300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaRed1200 = UIColor(named: "alphaRed1200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaRed1100 = UIColor(named: "alphaRed1100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaRed1000 = UIColor(named: "alphaRed1000", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaRed900 = UIColor(named: "alphaRed900", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaRed800 = UIColor(named: "alphaRed800", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaRed700 = UIColor(named: "alphaRed700", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaRed600 = UIColor(named: "alphaRed600", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaRed500 = UIColor(named: "alphaRed500", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaRed400 = UIColor(named: "alphaRed400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaRed300 = UIColor(named: "alphaRed300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaRed200 = UIColor(named: "alphaRed200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaRed100 = UIColor(named: "alphaRed100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGray1400 = UIColor(named: "alphaGray1400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGray1300 = UIColor(named: "alphaGray1300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGray1200 = UIColor(named: "alphaGray1200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGray1100 = UIColor(named: "alphaGray1100", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGray1000 = UIColor(named: "alphaGray1000", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGray900 = UIColor(named: "alphaGray900", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGray800 = UIColor(named: "alphaGray800", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGray700 = UIColor(named: "alphaGray700", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGray600 = UIColor(named: "alphaGray600", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGray500 = UIColor(named: "alphaGray500", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGray400 = UIColor(named: "alphaGray400", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGray300 = UIColor(named: "alphaGray300", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGray200 = UIColor(named: "alphaGray200", in: Bundle.module, compatibleWith: nil)!
-    public static let alphaGray100 = UIColor(named: "alphaGray100", in: Bundle.module, compatibleWith: nil)!
-    public static let pink1400 = UIColor(named: "pink1400", in: Bundle.module, compatibleWith: nil)!
-    public static let pink1300 = UIColor(named: "pink1300", in: Bundle.module, compatibleWith: nil)!
-    public static let pink1200 = UIColor(named: "pink1200", in: Bundle.module, compatibleWith: nil)!
-    public static let pink1100 = UIColor(named: "pink1100", in: Bundle.module, compatibleWith: nil)!
-    public static let pink1000 = UIColor(named: "pink1000", in: Bundle.module, compatibleWith: nil)!
-    public static let pink900 = UIColor(named: "pink900", in: Bundle.module, compatibleWith: nil)!
-    public static let pink800 = UIColor(named: "pink800", in: Bundle.module, compatibleWith: nil)!
-    public static let pink700 = UIColor(named: "pink700", in: Bundle.module, compatibleWith: nil)!
-    public static let pink600 = UIColor(named: "pink600", in: Bundle.module, compatibleWith: nil)!
-    public static let pink500 = UIColor(named: "pink500", in: Bundle.module, compatibleWith: nil)!
-    public static let pink400 = UIColor(named: "pink400", in: Bundle.module, compatibleWith: nil)!
-    public static let pink300 = UIColor(named: "pink300", in: Bundle.module, compatibleWith: nil)!
-    public static let pink200 = UIColor(named: "pink200", in: Bundle.module, compatibleWith: nil)!
-    public static let pink100 = UIColor(named: "pink100", in: Bundle.module, compatibleWith: nil)!
-    public static let fuchsia1400 = UIColor(named: "fuchsia1400", in: Bundle.module, compatibleWith: nil)!
-    public static let fuchsia1300 = UIColor(named: "fuchsia1300", in: Bundle.module, compatibleWith: nil)!
-    public static let fuchsia1200 = UIColor(named: "fuchsia1200", in: Bundle.module, compatibleWith: nil)!
-    public static let fuchsia1100 = UIColor(named: "fuchsia1100", in: Bundle.module, compatibleWith: nil)!
-    public static let fuchsia1000 = UIColor(named: "fuchsia1000", in: Bundle.module, compatibleWith: nil)!
-    public static let fuchsia900 = UIColor(named: "fuchsia900", in: Bundle.module, compatibleWith: nil)!
-    public static let fuchsia800 = UIColor(named: "fuchsia800", in: Bundle.module, compatibleWith: nil)!
-    public static let fuchsia700 = UIColor(named: "fuchsia700", in: Bundle.module, compatibleWith: nil)!
-    public static let fuchsia600 = UIColor(named: "fuchsia600", in: Bundle.module, compatibleWith: nil)!
-    public static let fuchsia500 = UIColor(named: "fuchsia500", in: Bundle.module, compatibleWith: nil)!
-    public static let fuchsia400 = UIColor(named: "fuchsia400", in: Bundle.module, compatibleWith: nil)!
-    public static let fuchsia300 = UIColor(named: "fuchsia300", in: Bundle.module, compatibleWith: nil)!
-    public static let fuchsia200 = UIColor(named: "fuchsia200", in: Bundle.module, compatibleWith: nil)!
-    public static let fuchsia100 = UIColor(named: "fuchsia100", in: Bundle.module, compatibleWith: nil)!
-    public static let purple1400 = UIColor(named: "purple1400", in: Bundle.module, compatibleWith: nil)!
-    public static let purple1300 = UIColor(named: "purple1300", in: Bundle.module, compatibleWith: nil)!
-    public static let purple1200 = UIColor(named: "purple1200", in: Bundle.module, compatibleWith: nil)!
-    public static let purple1100 = UIColor(named: "purple1100", in: Bundle.module, compatibleWith: nil)!
-    public static let purple1000 = UIColor(named: "purple1000", in: Bundle.module, compatibleWith: nil)!
-    public static let purple900 = UIColor(named: "purple900", in: Bundle.module, compatibleWith: nil)!
-    public static let purple800 = UIColor(named: "purple800", in: Bundle.module, compatibleWith: nil)!
-    public static let purple700 = UIColor(named: "purple700", in: Bundle.module, compatibleWith: nil)!
-    public static let purple600 = UIColor(named: "purple600", in: Bundle.module, compatibleWith: nil)!
-    public static let purple500 = UIColor(named: "purple500", in: Bundle.module, compatibleWith: nil)!
-    public static let purple400 = UIColor(named: "purple400", in: Bundle.module, compatibleWith: nil)!
-    public static let purple300 = UIColor(named: "purple300", in: Bundle.module, compatibleWith: nil)!
-    public static let purple200 = UIColor(named: "purple200", in: Bundle.module, compatibleWith: nil)!
-    public static let purple100 = UIColor(named: "purple100", in: Bundle.module, compatibleWith: nil)!
-    public static let blue1400 = UIColor(named: "blue1400", in: Bundle.module, compatibleWith: nil)!
-    public static let blue1300 = UIColor(named: "blue1300", in: Bundle.module, compatibleWith: nil)!
-    public static let blue1200 = UIColor(named: "blue1200", in: Bundle.module, compatibleWith: nil)!
-    public static let blue1100 = UIColor(named: "blue1100", in: Bundle.module, compatibleWith: nil)!
-    public static let blue1000 = UIColor(named: "blue1000", in: Bundle.module, compatibleWith: nil)!
-    public static let blue900 = UIColor(named: "blue900", in: Bundle.module, compatibleWith: nil)!
-    public static let blue800 = UIColor(named: "blue800", in: Bundle.module, compatibleWith: nil)!
-    public static let blue700 = UIColor(named: "blue700", in: Bundle.module, compatibleWith: nil)!
-    public static let blue600 = UIColor(named: "blue600", in: Bundle.module, compatibleWith: nil)!
-    public static let blue500 = UIColor(named: "blue500", in: Bundle.module, compatibleWith: nil)!
-    public static let blue400 = UIColor(named: "blue400", in: Bundle.module, compatibleWith: nil)!
-    public static let blue300 = UIColor(named: "blue300", in: Bundle.module, compatibleWith: nil)!
-    public static let blue200 = UIColor(named: "blue200", in: Bundle.module, compatibleWith: nil)!
-    public static let blue100 = UIColor(named: "blue100", in: Bundle.module, compatibleWith: nil)!
-    public static let cyan1400 = UIColor(named: "cyan1400", in: Bundle.module, compatibleWith: nil)!
-    public static let cyan1300 = UIColor(named: "cyan1300", in: Bundle.module, compatibleWith: nil)!
-    public static let cyan1200 = UIColor(named: "cyan1200", in: Bundle.module, compatibleWith: nil)!
-    public static let cyan1100 = UIColor(named: "cyan1100", in: Bundle.module, compatibleWith: nil)!
-    public static let cyan1000 = UIColor(named: "cyan1000", in: Bundle.module, compatibleWith: nil)!
-    public static let cyan900 = UIColor(named: "cyan900", in: Bundle.module, compatibleWith: nil)!
-    public static let cyan800 = UIColor(named: "cyan800", in: Bundle.module, compatibleWith: nil)!
-    public static let cyan700 = UIColor(named: "cyan700", in: Bundle.module, compatibleWith: nil)!
-    public static let cyan600 = UIColor(named: "cyan600", in: Bundle.module, compatibleWith: nil)!
-    public static let cyan500 = UIColor(named: "cyan500", in: Bundle.module, compatibleWith: nil)!
-    public static let cyan400 = UIColor(named: "cyan400", in: Bundle.module, compatibleWith: nil)!
-    public static let cyan300 = UIColor(named: "cyan300", in: Bundle.module, compatibleWith: nil)!
-    public static let cyan200 = UIColor(named: "cyan200", in: Bundle.module, compatibleWith: nil)!
-    public static let cyan100 = UIColor(named: "cyan100", in: Bundle.module, compatibleWith: nil)!
-    public static let green1400 = UIColor(named: "green1400", in: Bundle.module, compatibleWith: nil)!
-    public static let green1300 = UIColor(named: "green1300", in: Bundle.module, compatibleWith: nil)!
-    public static let green1200 = UIColor(named: "green1200", in: Bundle.module, compatibleWith: nil)!
-    public static let green1100 = UIColor(named: "green1100", in: Bundle.module, compatibleWith: nil)!
-    public static let green1000 = UIColor(named: "green1000", in: Bundle.module, compatibleWith: nil)!
-    public static let green900 = UIColor(named: "green900", in: Bundle.module, compatibleWith: nil)!
-    public static let green800 = UIColor(named: "green800", in: Bundle.module, compatibleWith: nil)!
-    public static let green700 = UIColor(named: "green700", in: Bundle.module, compatibleWith: nil)!
-    public static let green600 = UIColor(named: "green600", in: Bundle.module, compatibleWith: nil)!
-    public static let green500 = UIColor(named: "green500", in: Bundle.module, compatibleWith: nil)!
-    public static let green400 = UIColor(named: "green400", in: Bundle.module, compatibleWith: nil)!
-    public static let green300 = UIColor(named: "green300", in: Bundle.module, compatibleWith: nil)!
-    public static let green200 = UIColor(named: "green200", in: Bundle.module, compatibleWith: nil)!
-    public static let green100 = UIColor(named: "green100", in: Bundle.module, compatibleWith: nil)!
-    public static let lime1400 = UIColor(named: "lime1400", in: Bundle.module, compatibleWith: nil)!
-    public static let lime1300 = UIColor(named: "lime1300", in: Bundle.module, compatibleWith: nil)!
-    public static let lime1200 = UIColor(named: "lime1200", in: Bundle.module, compatibleWith: nil)!
-    public static let lime1100 = UIColor(named: "lime1100", in: Bundle.module, compatibleWith: nil)!
-    public static let lime1000 = UIColor(named: "lime1000", in: Bundle.module, compatibleWith: nil)!
-    public static let lime900 = UIColor(named: "lime900", in: Bundle.module, compatibleWith: nil)!
-    public static let lime800 = UIColor(named: "lime800", in: Bundle.module, compatibleWith: nil)!
-    public static let lime700 = UIColor(named: "lime700", in: Bundle.module, compatibleWith: nil)!
-    public static let lime600 = UIColor(named: "lime600", in: Bundle.module, compatibleWith: nil)!
-    public static let lime500 = UIColor(named: "lime500", in: Bundle.module, compatibleWith: nil)!
-    public static let lime400 = UIColor(named: "lime400", in: Bundle.module, compatibleWith: nil)!
-    public static let lime300 = UIColor(named: "lime300", in: Bundle.module, compatibleWith: nil)!
-    public static let lime200 = UIColor(named: "lime200", in: Bundle.module, compatibleWith: nil)!
-    public static let lime100 = UIColor(named: "lime100", in: Bundle.module, compatibleWith: nil)!
-    public static let yellow1400 = UIColor(named: "yellow1400", in: Bundle.module, compatibleWith: nil)!
-    public static let yellow1300 = UIColor(named: "yellow1300", in: Bundle.module, compatibleWith: nil)!
-    public static let yellow1200 = UIColor(named: "yellow1200", in: Bundle.module, compatibleWith: nil)!
-    public static let yellow1100 = UIColor(named: "yellow1100", in: Bundle.module, compatibleWith: nil)!
-    public static let yellow1000 = UIColor(named: "yellow1000", in: Bundle.module, compatibleWith: nil)!
-    public static let yellow900 = UIColor(named: "yellow900", in: Bundle.module, compatibleWith: nil)!
-    public static let yellow800 = UIColor(named: "yellow800", in: Bundle.module, compatibleWith: nil)!
-    public static let yellow700 = UIColor(named: "yellow700", in: Bundle.module, compatibleWith: nil)!
-    public static let yellow600 = UIColor(named: "yellow600", in: Bundle.module, compatibleWith: nil)!
-    public static let yellow500 = UIColor(named: "yellow500", in: Bundle.module, compatibleWith: nil)!
-    public static let yellow400 = UIColor(named: "yellow400", in: Bundle.module, compatibleWith: nil)!
-    public static let yellow300 = UIColor(named: "yellow300", in: Bundle.module, compatibleWith: nil)!
-    public static let yellow200 = UIColor(named: "yellow200", in: Bundle.module, compatibleWith: nil)!
-    public static let yellow100 = UIColor(named: "yellow100", in: Bundle.module, compatibleWith: nil)!
-    public static let orange1400 = UIColor(named: "orange1400", in: Bundle.module, compatibleWith: nil)!
-    public static let orange1300 = UIColor(named: "orange1300", in: Bundle.module, compatibleWith: nil)!
-    public static let orange1200 = UIColor(named: "orange1200", in: Bundle.module, compatibleWith: nil)!
-    public static let orange1100 = UIColor(named: "orange1100", in: Bundle.module, compatibleWith: nil)!
-    public static let orange1000 = UIColor(named: "orange1000", in: Bundle.module, compatibleWith: nil)!
-    public static let orange900 = UIColor(named: "orange900", in: Bundle.module, compatibleWith: nil)!
-    public static let orange800 = UIColor(named: "orange800", in: Bundle.module, compatibleWith: nil)!
-    public static let orange700 = UIColor(named: "orange700", in: Bundle.module, compatibleWith: nil)!
-    public static let orange600 = UIColor(named: "orange600", in: Bundle.module, compatibleWith: nil)!
-    public static let orange500 = UIColor(named: "orange500", in: Bundle.module, compatibleWith: nil)!
-    public static let orange400 = UIColor(named: "orange400", in: Bundle.module, compatibleWith: nil)!
-    public static let orange300 = UIColor(named: "orange300", in: Bundle.module, compatibleWith: nil)!
-    public static let orange200 = UIColor(named: "orange200", in: Bundle.module, compatibleWith: nil)!
-    public static let orange100 = UIColor(named: "orange100", in: Bundle.module, compatibleWith: nil)!
-    public static let red1400 = UIColor(named: "red1400", in: Bundle.module, compatibleWith: nil)!
-    public static let red1300 = UIColor(named: "red1300", in: Bundle.module, compatibleWith: nil)!
-    public static let red1200 = UIColor(named: "red1200", in: Bundle.module, compatibleWith: nil)!
-    public static let red1100 = UIColor(named: "red1100", in: Bundle.module, compatibleWith: nil)!
-    public static let red1000 = UIColor(named: "red1000", in: Bundle.module, compatibleWith: nil)!
-    public static let red900 = UIColor(named: "red900", in: Bundle.module, compatibleWith: nil)!
-    public static let red800 = UIColor(named: "red800", in: Bundle.module, compatibleWith: nil)!
-    public static let red700 = UIColor(named: "red700", in: Bundle.module, compatibleWith: nil)!
-    public static let red600 = UIColor(named: "red600", in: Bundle.module, compatibleWith: nil)!
-    public static let red500 = UIColor(named: "red500", in: Bundle.module, compatibleWith: nil)!
-    public static let red400 = UIColor(named: "red400", in: Bundle.module, compatibleWith: nil)!
-    public static let red300 = UIColor(named: "red300", in: Bundle.module, compatibleWith: nil)!
-    public static let red200 = UIColor(named: "red200", in: Bundle.module, compatibleWith: nil)!
-    public static let red100 = UIColor(named: "red100", in: Bundle.module, compatibleWith: nil)!
-    public static let gray1400 = UIColor(named: "gray1400", in: Bundle.module, compatibleWith: nil)!
-    public static let gray1300 = UIColor(named: "gray1300", in: Bundle.module, compatibleWith: nil)!
-    public static let gray1200 = UIColor(named: "gray1200", in: Bundle.module, compatibleWith: nil)!
-    public static let gray1100 = UIColor(named: "gray1100", in: Bundle.module, compatibleWith: nil)!
-    public static let gray1000 = UIColor(named: "gray1000", in: Bundle.module, compatibleWith: nil)!
-    public static let gray900 = UIColor(named: "gray900", in: Bundle.module, compatibleWith: nil)!
-    public static let gray800 = UIColor(named: "gray800", in: Bundle.module, compatibleWith: nil)!
-    public static let gray700 = UIColor(named: "gray700", in: Bundle.module, compatibleWith: nil)!
-    public static let gray600 = UIColor(named: "gray600", in: Bundle.module, compatibleWith: nil)!
-    public static let gray500 = UIColor(named: "gray500", in: Bundle.module, compatibleWith: nil)!
-    public static let gray400 = UIColor(named: "gray400", in: Bundle.module, compatibleWith: nil)!
-    public static let gray300 = UIColor(named: "gray300", in: Bundle.module, compatibleWith: nil)!
-    public static let gray200 = UIColor(named: "gray200", in: Bundle.module, compatibleWith: nil)!
-    public static let gray100 = UIColor(named: "gray100", in: Bundle.module, compatibleWith: nil)!
-    public static let themeBg = UIColor(named: "themeBg", in: Bundle.module, compatibleWith: nil)!
-    public static let iconOnSolidPrimary = themeBg
-    public static let iconInfoPrimary = blue900
-    public static let iconSuccessPrimary = green900
-    public static let iconCriticalPrimary = red900
-    public static let iconAccentPrimary = green900
-    public static let iconAccentTertiary = green800
-    public static let iconQuaternaryAlpha = alphaGray700
-    public static let iconTertiaryAlpha = alphaGray800
-    public static let iconSecondaryAlpha = alphaGray900
-    public static let iconPrimaryAlpha = alphaGray1400
-    public static let iconDisabled = gray700
-    public static let iconQuaternary = gray700
-    public static let iconTertiary = gray800
-    public static let iconSecondary = gray900
-    public static let iconPrimary = gray1400
-    public static let borderInfoSubtle = blue500
-    public static let borderSuccessSubtle = green500
-    public static let borderCriticalSubtle = red500
-    public static let borderCriticalHovered = red1000
-    public static let borderCriticalPrimary = red900
-    public static let borderInteractiveHovered = gray1100
-    public static let borderInteractiveSecondary = gray600
-    public static let borderInteractivePrimary = gray800
-    public static let borderFocused = blue900
-    public static let borderDisabled = gray500
-    public static let bgSubtleSecondaryLevel0 = themeBg
-    public static let bgAccentPressed = green1100
-    public static let bgAccentHovered = green1000
-    public static let bgAccentRest = green900
-    public static let bgDecorative6 = orange300
-    public static let bgDecorative5 = pink300
-    public static let bgDecorative4 = purple300
-    public static let bgDecorative3 = fuchsia300
-    public static let bgDecorative2 = cyan300
-    public static let bgDecorative1 = lime300
-    public static let bgInfoSubtle = blue200
-    public static let bgSuccessSubtle = green200
-    public static let bgCriticalSubtleHovered = red300
-    public static let bgCriticalSubtle = red200
-    public static let bgCriticalHovered = red1000
-    public static let bgCriticalPrimary = red900
-    public static let bgActionSecondaryPressed = alphaGray300
-    public static let bgActionSecondaryHovered = alphaGray200
-    public static let bgActionSecondaryRest = themeBg
-    public static let bgActionPrimaryDisabled = gray700
-    public static let bgActionPrimaryPressed = gray1100
-    public static let bgActionPrimaryHovered = gray1200
-    public static let bgActionPrimaryRest = gray1400
-    public static let bgCanvasDefaultLevel1 = gray300
-    public static let bgCanvasDisabled = gray200
-    public static let bgCanvasDefault = themeBg
-    public static let bgSubtleSecondary = gray300
-    public static let bgSubtlePrimary = gray400
-    public static let textDecorative6 = orange1100
-    public static let textDecorative5 = pink1100
-    public static let textDecorative4 = purple1100
-    public static let textDecorative3 = fuchsia1100
-    public static let textDecorative2 = cyan1100
-    public static let textDecorative1 = lime1100
-    public static let textOnSolidPrimary = themeBg
-    public static let textInfoPrimary = blue900
-    public static let textSuccessPrimary = green900
-    public static let textCriticalPrimary = red900
-    public static let textLinkExternal = blue900
-    public static let textActionAccent = green900
-    public static let textActionPrimary = gray1400
-    public static let textDisabled = gray800
-    public static let textPlaceholder = gray800
-    public static let textSecondary = gray900
-    public static let textPrimary = gray1400
+    public let iconOnSolidPrimary = CompoundCoreUIColorTokens.themeBg
+    public let iconInfoPrimary = CompoundCoreUIColorTokens.blue900
+    public let iconSuccessPrimary = CompoundCoreUIColorTokens.green900
+    public let iconCriticalPrimary = CompoundCoreUIColorTokens.red900
+    public let iconAccentPrimary = CompoundCoreUIColorTokens.green900
+    public let iconAccentTertiary = CompoundCoreUIColorTokens.green800
+    public let iconQuaternaryAlpha = CompoundCoreUIColorTokens.alphaGray700
+    public let iconTertiaryAlpha = CompoundCoreUIColorTokens.alphaGray800
+    public let iconSecondaryAlpha = CompoundCoreUIColorTokens.alphaGray900
+    public let iconPrimaryAlpha = CompoundCoreUIColorTokens.alphaGray1400
+    public let iconDisabled = CompoundCoreUIColorTokens.gray700
+    public let iconQuaternary = CompoundCoreUIColorTokens.gray700
+    public let iconTertiary = CompoundCoreUIColorTokens.gray800
+    public let iconSecondary = CompoundCoreUIColorTokens.gray900
+    public let iconPrimary = CompoundCoreUIColorTokens.gray1400
+    public let borderInfoSubtle = CompoundCoreUIColorTokens.blue500
+    public let borderSuccessSubtle = CompoundCoreUIColorTokens.green500
+    public let borderCriticalSubtle = CompoundCoreUIColorTokens.red500
+    public let borderCriticalHovered = CompoundCoreUIColorTokens.red1000
+    public let borderCriticalPrimary = CompoundCoreUIColorTokens.red900
+    public let borderInteractiveHovered = CompoundCoreUIColorTokens.gray1100
+    public let borderInteractiveSecondary = CompoundCoreUIColorTokens.gray600
+    public let borderInteractivePrimary = CompoundCoreUIColorTokens.gray800
+    public let borderFocused = CompoundCoreUIColorTokens.blue900
+    public let borderDisabled = CompoundCoreUIColorTokens.gray500
+    public let bgAccentPressed = CompoundCoreUIColorTokens.green1100
+    public let bgAccentHovered = CompoundCoreUIColorTokens.green1000
+    public let bgAccentRest = CompoundCoreUIColorTokens.green900
+    public let bgDecorative6 = CompoundCoreUIColorTokens.orange300
+    public let bgDecorative5 = CompoundCoreUIColorTokens.pink300
+    public let bgDecorative4 = CompoundCoreUIColorTokens.purple300
+    public let bgDecorative3 = CompoundCoreUIColorTokens.fuchsia300
+    public let bgDecorative2 = CompoundCoreUIColorTokens.cyan300
+    public let bgDecorative1 = CompoundCoreUIColorTokens.lime300
+    public let bgInfoSubtle = CompoundCoreUIColorTokens.blue200
+    public let bgSuccessSubtle = CompoundCoreUIColorTokens.green200
+    public let bgCriticalSubtleHovered = CompoundCoreUIColorTokens.red300
+    public let bgCriticalSubtle = CompoundCoreUIColorTokens.red200
+    public let bgCriticalHovered = CompoundCoreUIColorTokens.red1000
+    public let bgCriticalPrimary = CompoundCoreUIColorTokens.red900
+    public let bgActionSecondaryPressed = CompoundCoreUIColorTokens.alphaGray300
+    public let bgActionSecondaryHovered = CompoundCoreUIColorTokens.alphaGray200
+    public let bgActionSecondaryRest = CompoundCoreUIColorTokens.themeBg
+    public let bgActionPrimaryDisabled = CompoundCoreUIColorTokens.gray700
+    public let bgActionPrimaryPressed = CompoundCoreUIColorTokens.gray1100
+    public let bgActionPrimaryHovered = CompoundCoreUIColorTokens.gray1200
+    public let bgActionPrimaryRest = CompoundCoreUIColorTokens.gray1400
+    public let bgCanvasDisabled = CompoundCoreUIColorTokens.gray200
+    public let bgCanvasDefault = CompoundCoreUIColorTokens.themeBg
+    public let bgSubtleSecondary = CompoundCoreUIColorTokens.gray300
+    public let bgSubtlePrimary = CompoundCoreUIColorTokens.gray400
+    public let textDecorative6 = CompoundCoreUIColorTokens.orange1100
+    public let textDecorative5 = CompoundCoreUIColorTokens.pink1100
+    public let textDecorative4 = CompoundCoreUIColorTokens.purple1100
+    public let textDecorative3 = CompoundCoreUIColorTokens.fuchsia1100
+    public let textDecorative2 = CompoundCoreUIColorTokens.cyan1100
+    public let textDecorative1 = CompoundCoreUIColorTokens.lime1100
+    public let textOnSolidPrimary = CompoundCoreUIColorTokens.themeBg
+    public let textInfoPrimary = CompoundCoreUIColorTokens.blue900
+    public let textSuccessPrimary = CompoundCoreUIColorTokens.green900
+    public let textCriticalPrimary = CompoundCoreUIColorTokens.red900
+    public let textLinkExternal = CompoundCoreUIColorTokens.blue900
+    public let textActionAccent = CompoundCoreUIColorTokens.green900
+    public let textActionPrimary = CompoundCoreUIColorTokens.gray1400
+    public let textDisabled = CompoundCoreUIColorTokens.gray800
+    public let textPlaceholder = CompoundCoreUIColorTokens.gray800
+    public let textSecondary = CompoundCoreUIColorTokens.gray900
+    public let textPrimary = CompoundCoreUIColorTokens.gray1400
+    public init() { }
 }
